+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable user-facing changes to this project are documented here.
 
-## [Unreleased]
+## [1.6.0] — 2026-03-29
 
 ### New Features